rename to useRoyalty
